Added missing @can checks for Songs

+++ app/Http/Controllers/SongAttachmentController.php
@@ -11,11 +11,15 @@ class SongAttachmentController extends Controller
 {
     public function show(Song $song, SongAttachment $attachment)
     {
+        $this->authorize('view', $song);
+
         return response()->download( $attachment->path );
     }
 
     public function store(Song $song, SongAttachmentRequest $request): RedirectResponse
     {
+        $this->authorize('update', $song);
+
         $data = $request->validated();
 
         $attachment = SongAttachment::create([
@@ -30,6 +34,8 @@ class SongAttachmentController extends Controller
 
     public function destroy(Song $song, SongAttachment $attachment): RedirectResponse
     {
+        $this->authorize('update', $song);
+
         $attachment->delete();
 
         return redirect()->route('songs.show', [$song])->with(['status' => 'Attachment deleted. ', ]);
